Create CI/CD pipeline
This may be just the first of many commits for implementing CI/CD
pipeline with Github actions. There may be many small commits
subsequently, as we can't test the CI/CD pipeline locally and we'll
probably have many bugs. To debug, we can only commit and see how the
pipeline behaves.